WIP organize code structure for createform page
