+++ frontend/src/Pages/CreateFormJose/CreateFormJose.jsx
@@ -1,104 +1,12 @@
-import { useFieldArray, useForm } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import { useEffect, useState } from 'react';
 import { api } from '../../utils/api.js';
 import { useQuery, useQueryClient } from 'react-query';
 import { useNavigate, useParams, Link } from 'react-router-dom';
 import styles from './CreateForm.module.css'
+import FormForm from './FormForm.jsx';
+import Sidebar from '../../components/ui/Sidebar.jsx';
 
-const questionTypes = [{
-  value: "TextQuestion",
-  label: 'Text Question'
-}, {
-  value: "MultipleChoiceQuestion",
-  label: 'Multiple Choice Question'
-},
-{
-  value: "SingleChoiceQuestion",
-  label: 'Single Choice Question'
-}]
-
-const Choices = ({ register, control, index }) => {
-
-
-  const { fields, append, prepend, remove, swap, move, insert } = useFieldArray({
-    control,
-    name: `questions[${index}].choices`
-  });
-
-  return (
-    <div>
-      {fields.map((choice, choiceIndex) => (
-        <div className={styles.questionChoice} key={choice._id} >
-          <label>Choice</label>
-          <input type="text" {...register(`questions[${index}].choices[${choiceIndex}].label`)} />
-          <button type="button" onClick={() => remove(choiceIndex)}>x</button>
-        </div>
-      ))
-      }
-      <button type="button" onClick={() => append({})}>Add Choice</button>
-    </div >
-  )
-}
-
-const QuestionForm = ({ register, index, watch, control }) => {
-  const type = watch(`questions[${index}].type`);
-
-  return (
-    <div>
-      <label>Question Type</label>
-      <select {...register(`questions[${index}].type`, { value: "TextQuestion" })}>
-        {questionTypes.map((questionType, index) => <option value={questionType.value} key={index}>{questionType.label}</option>)}
-      </select>
-      <label>Text</label>
-      <input type="text" {...register(`questions[${index}].text`)} />
-      <label>Description</label>
-      <input type="text" {...register(`questions[${index}].description`)} />
-      {type !== "TextQuestion" && (
-        <Choices register={register} control={control} index={index} />
-      )}
-    </div>
-  )
-}
-
-const FormForm = ({ register, handleSubmit, onSubmit, watch, control }) => {
-
-  const { fields, append, prepend, remove, swap, move, insert } = useFieldArray({
-    control,
-    name: "questions"
-  });
-
-  return (<div>
-    <header className={styles.fixedHeader}>
-      <Link reloadDocument={false} className={styles.workspaceLink} to={'/workspace'}>my workspace</Link>
-
-      <form onSubmit={handleSubmit(onSubmit)} className={styles.question}>
-        <input id={styles.inputFormTitle} type="text" {...register('title')} />
-        <br />
-      </form>
-    </header>
-
-    <form onSubmit={handleSubmit(onSubmit)} className={styles.question}>
-
-      {fields.map((question, index) => (
-        <div className={styles.question} key={index}>
-          <QuestionForm
-            key={question.id}
-            register={register}
-            index={index}
-            watch={watch}
-            control={control}
-          />
-          <button type="button" onClick={() => remove(index)}>x</button>
-        </div>
-
-      ))}
-      <button type="button" onClick={() => append({})}>Add Question</button>
-      <button type="submit">Submit</button>
-
-    </form>
-  </div >)
-    ;
-}
 
 export const EditForm = () => {
   const { id } = useParams();
@@ -136,7 +44,10 @@ export const EditForm = () => {
   }
 
   return (
-    <FormForm register={register} control={control} handleSubmit={handleSubmit} onSubmit={onSubmit} watch={watch} />
+    <>
+
+      <FormForm register={register} control={control} handleSubmit={handleSubmit} onSubmit={onSubmit} watch={watch} />
+    </>
   )
 }
 
@@ -161,7 +72,6 @@ export const CreateForm = () => {
       console.log(response.data);
       queryClient.invalidateQueries('forms');
       navigate('/workspace');
-
     });
   }
